chore(lint): fix real-bug classes and silence test-architecture noise
- bugprone-implicit-widening-of-multiplication-result: explicit size_t
  casts in DFA allocation/realloc (search_tiers.c DFA_ENSURE macro),
  arena sizing, label capacity, SIMD table offsets, test bounds
- bugprone-suspicious-realloc-usage: temp-variable + NULL checks in
  CodeBuf growth (compiler_analysis, pattern_build, vm_exec) and the
  DFA final resize (search_tiers) so OOM never leaks the old buffer
- bugprone-narrowing-conversions: explicit (char) casts in the
  UPPER/LOWER template transforms (vm_exec, test_template_ops) and the
  UTF-8 byte literals in test_string_case
- bugprone-multi-level-implicit-pointer-conversion: explicit (void *)
  casts on snobol_free/realloc/free/memset arguments (28 sites)
- clang-diagnostic-missing-field-initializers: {nullptr} partial
  initializers -> {0} (core + tests)
- clang-diagnostic-unused-result: array-set calls in tests now assert
  their return values (+25 assertions)
- bugprone-unchecked-string-to-number-conversion: atoi -> strtol in
  test_runner and test_tables
- misc-use-internal-linkage: 6 core functions made static
  (pike_scan, tier_general_fallback, build_nfa_masks*,
  snobol_breakx_prescan, snobol_emit_arb); the check is then disabled —
  the remaining 177 hits are test-suite entry points that the runner
  exports by design
- misc-include-cleaner auto-fix across 73 files (one order-sensitive
  false positive in search_meta.c repaired by hand)
- misc-* one-offs: duplicate include, identical ternary branches,
  redundant casts/parens, misplaced widening cast, redundant-expression
  in tests
- .clang-tidy: disable bugprone-branch-clone (all 20 hits were
  fall-through cases / repeated failure handling) and
  misc-use-internal-linkage (test noise)
- Lint: 9,801 -> 385 warnings (96% down); full C suite 366 cases /
  73,325 assertions and PHP suite 568 / 1,344 green

+++ core/src/search_simd.c
@@ -593,7 +593,7 @@ static bool simd_nfa_exec_neon(const simd_nfa_t *nfa, const char *subject,
       /* Classify each byte: lookup table[byte] */
       uint8x16_t result = vdupq_n_u8(0);
       for (int g = 0; g < 16; g++) {
-        uint8x16_t sub_tbl = vld1q_u8(table + g * 16);
+        uint8x16_t sub_tbl = vld1q_u8(table + (size_t)g * 16);
         uint8x16_t tbl_res = vqtbl1q_u8(sub_tbl, lo);
         uint8x16_t mask = vceqq_u8(hi, vdupq_n_u8((uint8_t)g));
         result = vbslq_u8(mask, tbl_res, result);
@@ -621,7 +621,7 @@ static bool simd_nfa_exec_neon(const simd_nfa_t *nfa, const char *subject,
       uint8x16_t hi = vshrq_n_u8(data, 4);
       uint8x16_t result = vdupq_n_u8(0);
       for (int g = 0; g < 16; g++) {
-        uint8x16_t sub_tbl = vld1q_u8(table + g * 16);
+        uint8x16_t sub_tbl = vld1q_u8(table + (size_t)g * 16);
         uint8x16_t tbl_res = vqtbl1q_u8(sub_tbl, lo);
         uint8x16_t mask = vceqq_u8(hi, vdupq_n_u8((uint8_t)g));
         result = vbslq_u8(mask, tbl_res, result);
